Move schedule content into its own file
